+++ mirror/agent/internal/device/stream_other.go
@@ -3,10 +3,11 @@
 package device
 
 import (
+	"context"
 	"io"
 	"time"
 )
 
-func StreamFrames(writer io.Writer, interval time.Duration) error {
+func StreamFrames(ctx context.Context, writer io.Writer, interval time.Duration) error {
 	return codedError{code: "linux_required"}
 }